Add username to API Endpoint and Dynamic DB Path
Updated the API Endpoint to include the username field, as it was necessary for the user identification. Also, made changes in db path from hard coded to working directory based, to make it dynamic according to the system.

Updated several packages to their latest versions for improved performance and security. Made sure database queries include username for better user-message correlation.

The channel ID is now fetched from the query string of the URL, rather than the body of the request. This makes the API more intuitive and easier to use.

The commit includes also renaming for better consistency.